test(playwright): separate browser E2E tests from parser tests
Keep browser-dependent tests behind the e2e build tag while allowing parser tests to run in the standard test suite.
